provide an `history` protocol for autotrophic operators
An autotrophic operator is a data source, which currently
now how to answer to a `get` call (by virtue of its base
implementation), but has no way to answer to other keys
points of the api.

This commit provides a protocol for `history` (which opens
a path to the staircase also). We will see if it is needed
to provide more hooks of this kind in the future.