tsio: overhaul the series compatibility check
We were trying to check compatibility of all series metadata,
in a brain-dead way.

We now focus on tzawareness rather than whole metadata
(which is a can of worms but can wait a bit), that was the
real value of the existing checks anyway.

The naive switcher is now handled for all series using the expression
path and not just series names (not garanteed unique) or positions
(naive series then tzaware flipped by a 'naive call didn't work).