operators: accept new special parameters for the lisp env bindings
We have previously introduced top-level bindings for the query args.
We can now use them in operators that accepts dunder names in
positional parameters (and before regular parameters).

__interpreter__ was already there, but statically fixed at the start
of an interpretation sequence.

__from_value_date__, __to_value_date__ and __revision_date__ are now
available and will be fixed by the evaluator, as these values are
dynamically scoped (or "tree-scoped") and may get different values at
any point in the tree.

These parameters are omitted from the type-checking phase (as they are
not part of the user facing interface).

In the next steps, we will actually use the values.  We will show that
without them, the slice operator over-queries, and properly used, they
will help reduce the amount of querying (by using a smaller range when
needed).