operators/series: use the scoped parameters (__from_value_date__, __to_value_date__, __revision_date__)
Now, whenever a restricted scope has been put in place, the `series`
call will be effectively restricted.

However, we are not there yet, since we only have a top-level binding
for these variables.

We will introduce next a new notation to indicate whenever a new scope
is needed.

Note: in the evaluator, we switch from `quasiexpreval` to the full
blown evaluator, since we actually to handle (date "...") expressions
there.