scope: communicate to `series` the scope of any operator that wants to restrict it
We are introducing a new @argscope decorator for operators,
which registers a mapping from the operators boundaries
to the query boundaries.

At expand time we can now inject the restricted scope bindings
for good effect in the `series` operator.

As noted in the `series` operator NOTE, we have to perform a .loc on
the fetched series to have all tests remain green. That's because the
history interpreter pre-computes the histories (and currently these
precomputed histories haven't heard of the new args scope).

This will be addressed in subsequent commits.