operators/naive: define a scope to get a proper handling of the query boundaries
The `naive_transform` function is quite straightfowards: we use an
internal `tzaware-timestamp` lisp function to turn the naive query
dates into a tzaware date in the timezone provided to `naive`.

We also need to teach scalar_min/max to handle heterogenous dates
(tzaware vs naive) since we now have that kind of situation.

At the top-level, we now carefully inject the query dates as provided
(that is naive -> naive, tzaware -> tzaware) since we had been a bit
sloppy right there.