tsio,interpreter: allow to configure the concurrency level
And use that in tests immediately.

This inflates quite significantly the test count since all `tsh`
objects are now send twice (with concurrency = 1, then 16).

The conftest has to be slightly re-organized to handle this.

Having done that, we know the two branches of `evaluator.pevaluate`
are covered.

We have to fix again the no-concurrency path (the asyncfuncs were not
properly handled).