operators/asof: a working implementation for .history
Doing the optimised path for history will introduce unwanted
complexity code-wise. Also it is doubtful that people will use
.history much with `asof` expression in it (indeed .history is mainly
useful for the staircase operations, and asof is in practice an
analysis tool that competes with the staircase).

Hence we introduce an `iter_revisions` internal api point, which might
grow as a general api in the future, as iterating over versions is,
memory-wise, the right thing to do (for various purposes).