formula/expanded: add new parameters "shownames" that will allow to prune the tree
Shownames is a container (must implement __contains__)
of series names for which we want the expansion to happen.
Other nodes wont be expanded.

It is the dual of "stopnames".

Principle: starting from the root, the (named) nodes are
expanded, iff one of the showname belong to their descendants.

This will reduce the amount of series that must be cached
and the amount of computation.

With the following formula:

A-+-B--C
  +-D--E--F--G

and the shownames : ['E'],
the result should look like:

A-+-B
  +-D--E